Changed structure so that only one header must be included
